requirements.txt
fastapi
numpy
opencv-python
Pillow
python-multipart
torch
transformers
uvicorn
